refactor: rename superclass.rs to supertype.rs
