+++ lib/skc_hir/src/lib.rs
@@ -3,12 +3,12 @@ pub mod signature;
 mod signatures;
 mod sk_method;
 mod sk_type;
-mod superclass;
+mod supertype;
 pub use crate::signature::*;
 pub use crate::signatures::MethodSignatures;
 pub use crate::sk_method::{SkMethod, SkMethodBody, SkMethods};
 pub use crate::sk_type::{SkClass, SkModule, SkType, SkTypeBase, SkTypes, WTable};
-pub use crate::superclass::Superclass;
+pub use crate::supertype::Superclass;
 use serde::{Deserialize, Serialize};
 use shiika_ast::LocationSpan;
 use shiika_core::{names::*, ty, ty::*};
